third_party/WebKit/LayoutTests/external/wpt/html/semantics/document-metadata/the-base-element/base_href_invalid-expected.txt

This is a testharness.js-based test.
FAIL base element with unparseable href should have .href getter return attr valueassert_equals: expected "//test:test" but got ""
Harness: the test ran to completion.
